Refactor backend services: migrate to gamr_backend_api_service and remove legacy code
