+++ .dockerignore
@@ -1,0 +1,24 @@
+__pycache__
+*.pyc
+*.pyo
+*.pyd
+
+.venv
+env
+venv
+*.env
+
+build
+dist
+*.egg-info
+
+.vscode
+.idea
+
+*.log
+
+Dockerfile
+.dockerignore
+
+.pytest_cache
+.ruff_cache
