+++ gamr_backend_api_service.egg-info/SOURCES.txt
@@ -1,0 +1,36 @@
+README.md
+pyproject.toml
+gamr_backend_api_service/__init__.py
+gamr_backend_api_service/__main__.py
+gamr_backend_api_service/settings.py
+gamr_backend_api_service.egg-info/PKG-INFO
+gamr_backend_api_service.egg-info/SOURCES.txt
+gamr_backend_api_service.egg-info/dependency_links.txt
+gamr_backend_api_service.egg-info/requires.txt
+gamr_backend_api_service.egg-info/top_level.txt
+gamr_backend_api_service/api/__init__.py
+gamr_backend_api_service/api/factory.py
+gamr_backend_api_service/api/routes/__init__.py
+gamr_backend_api_service/api/routes/base.py
+gamr_backend_api_service/api/routes/exception_handlers.py
+gamr_backend_api_service/auth/__init__.py
+gamr_backend_api_service/auth/exceptions.py
+gamr_backend_api_service/auth/token.py
+gamr_backend_api_service/data/__init__.py
+gamr_backend_api_service/data/user_manager.py
+gamr_backend_api_service/database_connector/__init__.py
+gamr_backend_api_service/database_connector/abstract_database_connector.py
+gamr_backend_api_service/database_connector/firebase.py
+gamr_backend_api_service/model_services/__init__.py
+gamr_backend_api_service/model_services/abstract_model_service.py
+gamr_backend_api_service/model_services/hugging_face/__init__.py
+gamr_backend_api_service/model_services/hugging_face/errors.py
+gamr_backend_api_service/model_services/hugging_face/service.py
+gamr_backend_api_service/models/__init__.py
+gamr_backend_api_service/models/errors.py
+gamr_backend_api_service/models/image_data.py
+gamr_backend_api_service/models/image_response.py
+gamr_backend_api_service/models/user.py
+tests/test_api.py
+tests/test_auth.py
+tests/test_validation.py
