+++ gamr_backend_api_service.egg-info/requires.txt
@@ -1,0 +1,8 @@
+fastapi==0.115.0
+uvicorn==0.34.0
+numpy==1.26.4
+gradio-client==1.5.0
+python-multipart==0.0.18
+pyjwt==2.10.1
+firebase-admin==6.6.0
+pydantic-settings
