+++ gamr_backend_api_service.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+gamr_backend_api_service
