+++ gamr_backend_api_service/auth/__init__.py
@@ -1,4 +1,4 @@
-from gamr_backend_service.models.errors import UserBlank
+from gamr_backend_api_service.models.errors import UserBlank
 
 from .exceptions import TokenNotDecoded, UserNotExists
 from .token import TokenGenerator
