+++ gamr_backend_api_service/auth/exceptions.py
