+++ gamr_backend_api_service/data/user_manager.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass, field
 
-from gamr_backend_service.auth.exceptions import (
+from gamr_backend_api_service.auth.exceptions import (
     UserNotExists,
 )
-from gamr_backend_service.database_connector import AbstractDatabaseConnector, FirestoreConnector
-from gamr_backend_service.models.user import User
+from gamr_backend_api_service.database_connector import AbstractDatabaseConnector, FirestoreConnector
+from gamr_backend_api_service.models.user import User
 
 
 @dataclass
