+++ gamr_backend_api_service/database_connector/abstract_database_connector.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 
-from gamr_backend_service.models.user import User
+from gamr_backend_api_service.models.user import User
 
 
 @dataclass
