+++ gamr_backend_api_service/model_services/abstract_model_service.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 
-from gamr_backend_service.models.image_data import ImageData
+from gamr_backend_api_service.models.image_data import ImageData
 
 
 @dataclass
