+++ gamr_backend_api_service/models/image_response.py
