+++ pyproject.toml
@@ -1,0 +1,29 @@
+[project]
+name = "gamr_backend_api_service"
+version = "1.0.0"
+description = "A FastAPI backend for image recognition"
+readme = "README.md"
+requires-python = ">=3.12,<4.0"
+dependencies = [
+    "fastapi==0.115.0",
+    "uvicorn==0.34.0",
+    "numpy==1.26.4",
+    "gradio-client==1.5.0",
+    "python-multipart==0.0.18",
+    "pyjwt==2.10.1",
+    "firebase-admin==6.6.0",
+    "pydantic-settings",
+]
+
+[dependency-groups]
+dev = [
+    "pytest==8.3.4",
+    "pytest-cov==6.0.0",
+    "flake8==7.1.2",
+    "black==25.1.0",
+    "pre-commit==4.1.0",
+    "ipykernel",
+    "pillow==11.1.0",
+    "poetry==2.1.1",
+    "uv==0.6.3",
+]
